auth, appointments & members overview

+++ lib/widgets/rhst_spacer.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
 import 'package:rhst/constants.dart';
 
 class RHSTSpacer extends StatelessWidget {
-  final int multiplier;
+  final double multiplier;
   const RHSTSpacer(this.multiplier, {Key? key}) : super(key: key);
 
   @override
